SpatialQueryManager -> QueryManager
